requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.5.2
requests==2.22.0
requests-toolbelt==0.9.1
slackclient==2.3.1
urllib3==1.25.7
